+++ src/main/java/com/skilles/cannacraft/blocks/weedCrop/WeedCropEntity.java
@@ -4,8 +4,9 @@ import com.skilles.cannacraft.registry.ModConfig;
 import com.skilles.cannacraft.registry.ModEntities;
 import com.skilles.cannacraft.strain.Gene;
 import com.skilles.cannacraft.strain.GeneTypes;
-import com.skilles.cannacraft.strain.GeneticsManager;
 import com.skilles.cannacraft.strain.StrainMap;
+import com.skilles.cannacraft.util.CrossUtil;
+import com.skilles.cannacraft.util.MiscUtil;
 import me.shedaniel.autoconfig.AutoConfig;
 import me.shedaniel.autoconfig.ConfigData;
 import net.fabricmc.fabric.api.block.entity.BlockEntityClientSerializable;
@@ -24,8 +25,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Random;
 
-import static com.skilles.cannacraft.blocks.weedCrop.WeedCrop.withBreeding;
-import static com.skilles.cannacraft.strain.StrainMap.*;
+import static com.skilles.cannacraft.util.StrainUtil.*;
 
 // TODO: drop seedId, seedThc if male
 // TODO: cross plants by collecting pollen from male
@@ -67,7 +67,6 @@ public class WeedCropEntity extends BlockEntity implements BlockEntityClientSeri
         float multiplier = 1;
         if(hasGene(GeneTypes.SPEED)) multiplier *= (1.0F + ((float) getGene(GeneTypes.SPEED).level()) / 2.0F); // 1: 50%, 2: 100%, 3: 150%
         multiplier *= config.getSpeed();
-        System.out.println(multiplier);
         return multiplier;
     }
 
@@ -75,9 +74,9 @@ public class WeedCropEntity extends BlockEntity implements BlockEntityClientSeri
         breedTimer = 1;
     }
     public void stopBreeding() {
-        breedCrops(getWorld(), pos, GeneticsManager.random());
+        breedCrops(getWorld(), pos, MiscUtil.random());
         breedTimer = -1;
-        getWorld().setBlockState(pos, withBreeding(getCachedState(), false));
+        getWorld().setBlockState(pos, WeedCrop.withBreeding(getCachedState(), false));
     }
     public int breedingProgress() {
         return breedTimer;
@@ -127,7 +126,7 @@ public class WeedCropEntity extends BlockEntity implements BlockEntityClientSeri
         }
     }
     public boolean hasGene(GeneTypes gene) {
-        return GeneticsManager.NbtListContains(this.attributes, gene.getName());
+        return MiscUtil.NbtListContains(this.attributes, gene.getName());
     }
     public Gene getGene(GeneTypes type) {
         if(!this.attributes.isEmpty()) {
@@ -178,10 +177,10 @@ public class WeedCropEntity extends BlockEntity implements BlockEntityClientSeri
             }
             if (maleId == 0) {
                 thc = ogTag.getInt("Seed THC");
-                maleId = stringArray.size() > 1 ? indexOf((stringArray.get(GeneticsManager.random().nextInt(stringArray.size() - 1)))) : indexOf(stringArray.get(0));
+                maleId = stringArray.size() > 1 ? indexOf((stringArray.get(MiscUtil.random().nextInt(stringArray.size() - 1)))) : indexOf(stringArray.get(0));
             }
             // Set thc
-            ogTag.putInt("Seed THC", GeneticsManager.crossThc(thc, ogTag.getInt("THC")));
+            ogTag.putInt("Seed THC", CrossUtil.crossThc(thc, ogTag.getInt("THC")));
             System.out.println("THC: " + ogTag.getInt("Seed THC"));
             // Set name/type
             int randId = random.nextInt(stringArray.size());
@@ -189,8 +188,8 @@ public class WeedCropEntity extends BlockEntity implements BlockEntityClientSeri
             String name2 = getStrain(maleId).name();
             StrainMap.Type type1 = getStrain(id).type();
             StrainMap.Type type2 = typeArray.get(randId);
-            String crossedName = GeneticsManager.crossStrains(name1, name2);
-            if (!isPresent(crossedName)) addStrain(crossedName, GeneticsManager.crossTypes(type1, type2));
+            String crossedName = CrossUtil.crossStrains(name1, name2);
+            if (!isPresent(crossedName)) addStrain(crossedName, CrossUtil.crossTypes(type1, type2));
             ogTag.putInt("Seed ID", indexOf(crossedName));
 
             // Save nbt
