+++ src/main/java/com/skilles/cannacraft/items/WeedFruit.java
@@ -1,7 +1,8 @@
 package com.skilles.cannacraft.items;
 
-import com.skilles.cannacraft.strain.GeneticsManager;
 import com.skilles.cannacraft.strain.StrainMap;
+import com.skilles.cannacraft.util.MiscUtil;
+import com.skilles.cannacraft.util.StrainUtil;
 import net.minecraft.client.item.TooltipContext;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
@@ -19,8 +20,8 @@ public class WeedFruit extends Item {
     public Text getName(ItemStack stack) {
         if (stack.hasTag()) {
             NbtCompound tag = stack.getSubTag("cannacraft:strain");
-            if(StrainMap.getStrain(tag.getInt("ID")).type().equals(StrainMap.Type.UNKNOWN)) tag.putInt("ID", 0);
-            return tag.getBoolean("Identified") ? Text.of(StrainMap.getStrain(tag.getInt("ID")).name()) : Text.of("Unidentified Cannabis");
+            if(StrainUtil.getStrain(tag.getInt("ID")).type().equals(StrainMap.Type.UNKNOWN)) tag.putInt("ID", 0);
+            return tag.getBoolean("Identified") ? Text.of(StrainUtil.getStrain(tag.getInt("ID")).name()) : Text.of("Unidentified Cannabis");
         }
         return super.getName(stack);
     }
@@ -30,7 +31,7 @@ public class WeedFruit extends Item {
         super.appendTooltip(stack, world, tooltip, context);
         NbtCompound tag = stack.getOrCreateSubTag("cannacraft:strain");
         if (tag != null && tag.contains("ID") && !(tag.getInt("ID") == 0)) { // checks if ID is set to actual strain
-            GeneticsManager.appendTooltips(tooltip, tag);
+            MiscUtil.appendTooltips(tooltip, tag);
         }
     }
 }
