+++ src/main/java/com/skilles/cannacraft/items/WeedSeed.java
@@ -5,8 +5,9 @@ import com.skilles.cannacraft.blocks.machines.strainAnalyzer.StrainAnalyzerEntit
 import com.skilles.cannacraft.blocks.weedCrop.WeedCrop;
 import com.skilles.cannacraft.components.StrainInterface;
 import com.skilles.cannacraft.registry.ModMisc;
-import com.skilles.cannacraft.strain.GeneticsManager;
 import com.skilles.cannacraft.strain.StrainMap;
+import com.skilles.cannacraft.util.MiscUtil;
+import com.skilles.cannacraft.util.StrainUtil;
 import net.minecraft.block.Block;
 import net.minecraft.block.FarmlandBlock;
 import net.minecraft.block.entity.BlockEntity;
@@ -35,8 +36,8 @@ public class WeedSeed extends AliasedBlockItem {
     public Text getName(ItemStack stack) {
         if (stack.hasTag()) {
             NbtCompound tag = stack.getSubTag("cannacraft:strain");
-            if(StrainMap.getStrain(tag.getInt("ID")).type().equals(StrainMap.Type.UNKNOWN)) tag.putInt("ID", 0);
-            return tag.getBoolean("Identified") ? Text.of(StrainMap.getStrain(tag.getInt("ID")).name() + " Seeds") : Text.of("Unidentified Seeds");
+            if(StrainUtil.getStrain(tag.getInt("ID")).type().equals(StrainMap.Type.UNKNOWN)) tag.putInt("ID", 0);
+            return tag.getBoolean("Identified") ? Text.of(StrainUtil.getStrain(tag.getInt("ID")).name() + " Seeds") : Text.of("Unidentified Seeds");
         }
         return super.getName(stack);
     }
@@ -72,7 +73,7 @@ public class WeedSeed extends AliasedBlockItem {
                 BlockEntity blockEntity = context.getWorld().getBlockEntity(context.getBlockPos());
                 NbtCompound tag = blockEntity.writeNbt(new NbtCompound());
                 if (context.getPlayer().isSneaking()) {
-                    System.out.println("Strain of crop: " + StrainMap.getStrain(tag.getInt("ID")).name()
+                    System.out.println("Strain of crop: " + StrainUtil.getStrain(tag.getInt("ID")).name()
                             + " Identified: " + tag.getBoolean("identified")
                             + " THC: " + tag.getInt("THC"));
                     System.out.println(tag);
@@ -94,7 +95,7 @@ public class WeedSeed extends AliasedBlockItem {
         super.appendTooltip(stack, world, tooltip, context);
         NbtCompound tag = stack.getSubTag("cannacraft:strain");
         if (tag != null && tag.contains("ID") && !(tag.getInt("ID") == 0)) { // checks if ID is set to actual strain
-            GeneticsManager.appendTooltips(tooltip, tag);
+            MiscUtil.appendTooltips(tooltip, tag);
         }
