+++ src/main/java/com/skilles/cannacraft/misc/HighEffect.java
@@ -2,7 +2,8 @@ package com.skilles.cannacraft.misc;
 
 import com.skilles.cannacraft.mixins.StatusEffectAccessor;
 import com.skilles.cannacraft.registry.ModMisc;
-import com.skilles.cannacraft.strain.GeneticsManager;
+import com.skilles.cannacraft.util.HighUtil;
+import com.skilles.cannacraft.util.MiscUtil;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.attribute.AttributeContainer;
 import net.minecraft.entity.effect.StatusEffect;
@@ -27,10 +28,10 @@ public class HighEffect extends StatusEffect {
             PlayerEntity player = (PlayerEntity) entity;
             StatusEffectInstance currentEffect = entity.getStatusEffect(this);
             if(currentEffect.getDuration() % 600 == 0) {
-                ((StatusEffectAccessor) currentEffect).setAmplifier(GeneticsManager.durationToAmplifier(currentEffect.getDuration()));
+                ((StatusEffectAccessor) currentEffect).setAmplifier(HighUtil.durationToAmplifier(currentEffect.getDuration()));
             } else if(currentEffect.getAmplifier() >= 1) {
-                for (int i = 0; i < GeneticsManager.random().nextInt(7) * currentEffect.getAmplifier(); i++) {
-                    player.updateTrackedPositionAndAngles(player.getX(), player.getY(), player.getZ(), player.bodyYaw + (GeneticsManager.random().nextInt(20) - 10) * currentEffect.getAmplifier(), player.getPitch() + GeneticsManager.random().nextInt(5) * currentEffect.getAmplifier(), 150, true);
+                for (int i = 0; i < MiscUtil.random().nextInt(7) * currentEffect.getAmplifier(); i++) {
+                    player.updateTrackedPositionAndAngles(player.getX(), player.getY(), player.getZ(), player.bodyYaw + (MiscUtil.random().nextInt(20) - 10) * currentEffect.getAmplifier(), player.getPitch() + MiscUtil.random().nextInt(5) * currentEffect.getAmplifier(), 150, true);
                 }
             }
         }
