+++ src/main/java/com/skilles/cannacraft/mixins/ItemEntityMixin.java
@@ -1,8 +1,8 @@
 package com.skilles.cannacraft.mixins;
 
 import com.skilles.cannacraft.registry.ModItems;
-import com.skilles.cannacraft.strain.GeneticsManager;
-import com.skilles.cannacraft.strain.StrainMap;
+import com.skilles.cannacraft.util.MiscUtil;
+import com.skilles.cannacraft.util.StrainUtil;
 import net.minecraft.entity.ItemEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NbtCompound;
@@ -28,8 +28,8 @@ public abstract class ItemEntityMixin {
         if(stack.getItem().equals(ModItems.WEED_SEED) || stack.getItem().equals(ModItems.WEED_FRUIT)) {
             NbtCompound tag = stack.getSubTag("cannacraft:strain");
             if(tag != null && tag.contains("ID")) {
-                if(tag.getInt("ID") == 0) GeneticsManager.randomizeTag(tag);
-                if(!tag.contains("THC") || (tag.contains("THC") && tag.getInt("THC") < 13)) tag.putInt("THC", StrainMap.normalDist(15, 5, 13));
+                if(tag.getInt("ID") == 0) MiscUtil.randomizeTag(tag);
+                if(!tag.contains("THC") || (tag.contains("THC") && tag.getInt("THC") < 13)) tag.putInt("THC", StrainUtil.normalDist(15, 5, 13));
             }
         }
     }
