+++ src/main/java/com/skilles/cannacraft/strain/StrainMap.java
@@ -5,6 +5,7 @@ import com.google.common.collect.HashBiMap;
 import com.google.common.reflect.TypeToken;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
+import com.skilles.cannacraft.util.StrainUtil;
 
 import java.io.FileWriter;
 import java.io.Reader;
@@ -13,7 +14,6 @@ import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.Random;
 /**
  * This class is responsible for a lot of strain related data management
  */
@@ -39,122 +39,7 @@ public final class StrainMap {
         //GeneticsManager.test();
         save();
     }
-    public static int getStrainCount() {
-        return strainArray.size();
-    }
-    public static Strain getStrain(int index) {
-        if(!strainArray.containsKey(index)) return strainArray.get(0);
-        return strainArray.get(index);
-    }
-    public static Strain getStrain(String name) {
-        if(!strainList.containsKey(name)) return strainArray.get(0);
-        return strainList.get(name);
-    }
-    public static int indexOf(Strain strain) {
-        return strainArray.inverse().get(strain);
-    }
-    public static int indexOf(String name) {
-        if(!strainList.containsKey(name)) return 0;
-        return indexOf(toStrain(name));
-    }
-    public static void addStrain(String name, Type type) {
-        Strain strain = new Strain(name, type);
-        if (!type.equals(Type.UNKNOWN)) {
-            int index;
 
-            index = strainArray.size();
-
-            if (strainList.containsKey(name)) {
-                strain = null;
-                System.out.println("No duplicate strains!");
-            } else {
-                strainArray.put(index, strain);
-                strainList.put(strainArray.get(index).name(), strainArray.get(index));
-                save();
-            }
-        } else {
-            strainArray.put(0, strain);
-        }
-    }
-
-    /**
-     * @param type the type to find
-     * @return returns true if the type is present in the strain list, false otherwise
-     */
-    private static boolean containsType(Type type) {
-        for (int i = 1; i < strainArray.size(); i++) {
-            if(strainArray.get(i).type().equals(type)) {
-                return true;
-            }
-        }
-        return false;
-    }
-
-    /**
-     * @param type the type to find
-     * @return returns 0 if that type is not found or the index of the first found (excluding index 0)
-     */
-    private static int indexOfType(Type type) {
-        for (int i = 1; i < strainArray.size(); i++) {
-            if(strainArray.get(i).type().equals(type)) {
-                return i;
-            }
-        }
-        return 0;
-    }
-    public static void removeStrain(int index) {
-        if(strainArray.containsKey(index)) {
-            /*boolean cascade = false;
-            strainList.remove(strainArray.get(index).name());
-            for (int i = index; strainArray.size() - 1 > index; i++) {
-                strainArray.forcePut(i, strainArray.get(i + 1));
-                strainArray.remove(i + 1);
-                cascade = true;
-            }
-            if (!cascade) {
-                strainArray.remove(index);
-            }*/
-            strainList.remove(strainArray.get(index).name());
-            strainArray.remove(index);
-        }
-    }
-    public static void resetStrains() {
-        strainArray.clear();
-        strainList.clear();
-        ogStrainCount = 4;
-        addStrain("Unknown", Type.UNKNOWN);
-        addStrain("OG Kush", Type.HYBRID);
-        addStrain("Purple Punch", Type.INDICA);
-        addStrain("Chem Trix", Type.SATIVA);
-        for (int i = 0; strainArray.size() > i; i++) {
-            System.out.println("Strain: "+ strainArray.get(i));
-            strainList.put(strainArray.get(i).name(),strainArray.get(i));
-            ogStrainCount++;
-        }
-    }
-    public static boolean isPresent(Strain strain) {
-        return strainArray.containsValue(strain);
-    }
-    public static boolean isPresent(String name) {
-        return strainList.containsKey(name);
-    }
-    public static int normalDist(int mean, int std, int min) {
-        Random random = new Random();
-        int newThc = (int) Math.round(random.nextGaussian()*std+mean);
-        if(newThc < min) {
-            newThc = min;
-        }
-        return newThc;
-    }
-    public static Strain toStrain(String name) {
-        return strainList.get(name);
-    }
-    public static Map<String, Strain> getNames() {
-        return strainList;
-    }
-    public static Map<Integer, Strain> getStrains() {
-        return strainArray;
-    }
     public static void save() {
         try (Writer writer = new FileWriter("strains.json")) {
             gson.toJson(strainArray, writer);
@@ -176,7 +61,7 @@ public final class StrainMap {
             }
         } catch(Exception e) {
             System.out.println("Error loading strains");
-            resetStrains();
+            StrainUtil.resetStrains();
             save();
         }
     }
